+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_start/setup.js
@@ -510,6 +510,7 @@ if (!gwaioSetupLoaded) {
             {
               // Custom
               customDifficulty: true,
+              difficultyName: "!LOC:Custom",
             },
           ];